recipes/requirements.txt
absl-py==0.13.0
argon2-cffi==20.1.0
asgiref==3.5.0
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==3.3.1
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
comtypes==1.1.11
coverage==6.3.2
cryptography==37.0.2
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.50.0
django-cleanup==6.0.0
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
entrypoints==0.3
enum34==1.1.10
et-xmlfile==1.1.0
Faker==12.0.1
ffmpeg==1.4
Flask==2.1.1
flatbuffers==1.12
fsspec==2022.3.0
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.34.1
gTTS==2.2.4
gym==0.7.0
h11==0.13.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.8.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mixer==7.2.2
mutagen==1.45.1
Naked==0.1.31
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
plaidbench==0.7.0
plaidml==0.7.0
plaidml-keras==0.7.0
pluggy==1.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.14.1
pygame==2.1.2
pyglet==1.5.21
Pygments==2.9.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyrsistent==0.18.0
PySocks==1.7.1
pyTelegramBotAPI==4.4.0
pytest==7.1.1
pytest-cov==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pyttsx3==2.90
pytube==12.0.0
pytz==2021.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.0
seaborn==0.11.2
selenium==4.1.5
Send2Trash==1.7.1
signals==0.0.2
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.35
sqlparse==0.4.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
tomli==2.0.1
tornado==6.1
tqdm==4.63.1
traitlets==5.0.5
trio==0.20.0
trio-websocket==0.9.2
txt==2020.11.12
typing-extensions==3.7.4.3
tzdata==2022.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.1.0
yt-dlp==2022.2.4
zipp==3.8.0
